(MODULES-6708) Install CA Certs for module downloading
Previously the CA Certs were not required to be explicitly installed however
the forge appears to be using newer certificates now which are not in the
default OpenSSL pack on Windows nodes.  This commit uses the install_ca_certs
helper to install the CA Certs prior to downloading modules from the forge.